Move ThingRecord back into RepositoryTest
